# swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "3.0"
  title: Trade-Server for CoinEx Chain
  description: A REST interface for state queries
...
schemes:
  - https
  - http
host: localhost:1317
securityDefinitions:
  kms:
    type: basic
paths:
  /market/tickers:
    get:
      tags:
        - Market
      summary: Query market tickers
      description: Query tickers info
      operationId: queryTickers
      produces:
        - application/json
      parameters:
        - in: query
          name: market_list
          description: Market count limited to 1~100
          required: true
          type: array
          items:
            type: string
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Tickers"
        500:
          description: Server internal error
  /misc/block-times:
    get:
      tags:
        - Misc
      summary: Query block time
      description: Query the block time up to given height
      operationId: queryBlockTimes
      produces:
        - application/json
      parameters:
        - in: query
          name: height
          description: Block height
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              type: integer
              format: int64
              example: 1565740800
              description: Unix timestamp
        500:
          description: Server internal error
  /market/depths:
    get:
      tags:
        - Market
      summary: Query market depth
      description: Query purchases and sales of a market at all price levels
      operationId: queryDepth
      produces:
        - application/json
      parameters:
        - in: query
          name: market
          description: stock/money
          required: true
          type: string
        - in: query
          name: count
          description: Querier count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              asks:
                type: array
                items:
                  $ref: "#/definitions/PricePoint"
              bids:
                type: array
                items:
                  $ref: "#/definitions/PricePoint"
        500:
          description: Server internal error
  /market/candle-sticks:
    get:
      tags:
        - Market
      summary: Query market candleStick
      description: Query candleStick until to given time
      operationId: queryCandleStick
      produces:
        - application/json
      parameters:
        - in: query
          name: market
          description: stock/money
          required: true
          type: string
        - in: query
          name: timespan
          description: Minute:16/Hour:32/Day:48
          required: true
          type: integer
          format: byte
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier candleStick count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/CandleStick'
        500:
          description: Server internal error
  /market/user-orders:
    get:
      tags:
        - Market
      summary: Query account's order
      description: Query account's order activities in all markets until to given time
      operationId: queryOrder
      produces:
        - application/json
      parameters:
        - in: query
          name: account
          description: Bech32 address
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier order count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/UserOrder'
        500:
          description: Server internal error
  /market/deals:
    get:
      tags:
        - Market
      summary: Query market deal
      description: Query market deal until to given time
      operationId: queryDeal
      produces:
        - application/json
      parameters:
        - in: query
          name: market
          description: stock/money
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier deal count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              deals:
                type: array
                items:
                  $ref: '#/definitions/FillOrderInfo'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
  /bancorlite/infos:
    get:
      tags:
        - Bancor
      summary: Query bancor info
      description: Query bancor info until to given time
      operationId: queryBancorInfo
      produces:
        - application/json
      parameters:
        - in: query
          name: market
          description: stock/money
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              bancor_info:
                type: array
                items:
                  $ref: '#/definitions/BancorInfo'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
  /bancorlite/trades:
    get:
      tags:
        - Bancor
      summary: Query bancor trade
      description: Query bancor trade until to given time
      operationId: queryBancorTrade
      produces:
        - application/json
      parameters:
        - in: query
          name: account
          description: Bech32 address
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              bancor_trade:
                type: array
                items:
                  $ref: '#/definitions/BancorTrade'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
  /expiry/redelegations:
    get:
      tags:
        - Expiry
      summary: Query redelegation
      description: Query delegator's redelegation-completion info
      operationId: queryRedelegation
      produces:
        - application/json
      parameters:
        - in: query
          name: account
          description: Bech32 address
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              redelegation:
                type: array
                items:
                  $ref: '#/definitions/Redelegation'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
  /expiry/unbondings:
    get:
      tags:
        - Expiry
      summary: Query Unbonding
      description: Query delegator's unbonding-completion info
      operationId: queryUnbonding
      produces:
        - application/json
      parameters:
        - in: query
          name: account
          description: Bech32 address
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              unbonding:
                type: array
                items:
                  $ref: '#/definitions/Unbonding'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
  /expiry/unlocks:
    get:
      tags:
        - Expiry
      summary: Query Unlock
      description: Query Unlock info
      operationId: queryUnlock
      produces:
        - application/json
      parameters:
        - in: query
          name: account
          description: Bech32 address
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              unlock:
                type: array
                items:
                  $ref: '#/definitions/Unlock'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
  /tx/incomes:
    get:
      tags:
        - Tx
      summary: Query account all income until to given time
      description: Query income info
      operationId: queryIncome
      produces:
        - application/json
      parameters:
        - in: query
          name: account
          description: Bech32 address
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              tx:
                type: array
                items:
                  $ref: '#/definitions/Tx'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
  /tx/txs:
    get:
      tags:
        - Tx
      summary: Query transactions
      description: Query transactions signed by given account until to given time
      operationId: queryTx
      produces:
        - application/json
      parameters:
        - in: query
          name: account
          description: Bech32 address
          required: true
          type: string
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              tx:
                type: array
                items:
                  $ref: '#/definitions/Tx'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
...
  /slash/slashings:
    get:
      tags:
        - Slash
      summary: Query validator slash info
      description: Query validator slash power, reason, and jailed status etc
      operationId: querySlash
      produces:
        - application/json
      parameters:
        - in: query
          name: time
          description: Unix timestamp
          required: true
          type: integer
          format: int64
        - in: query
          name: sid
          description: Sequence id
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          description: Querier count limited to 1024
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              slash:
                type: array
                items:
                  $ref: '#/definitions/Slash'
              timesid:
                type: array
                items:
                  type: integer
                  format: int64
                  description: Timesid entry are in pairs, a count correspond two entry(time,sid)
        500:
          description: Server internal error
definitions:
...
  Tickers:
    type: object
    properties:
      market:
        type: string
        example: "abc/cet"
      new:
        type: string
        example: "0"
        description: Newest price
      old:
        type: string
        example: "0"
        description: Old price one day ago
  PricePoint:
    type: object
    properties:
      price:
        type: string
        example: "0"
        description: Price of depth without precision consolidation
      amount:
        type: string
        example: "0"
        description: Amount of all the unfilled orders
  CandleStick:
    type: object
    properties:
      open:
        type: string
        example: "0"
        description: open price
      close:
        type: string
        example: "0"
        description: close price
      high:
        type: string
        example: "0"
        description: high price
      low:
        type: string
        example: "0"
        description: low price
      total:
        type: string
        example: "0"
        description: total deal
      unix_time:
        type: integer
        format: int64
        example: 1565740800
        description: ending unix time
      time_span:
        type: integer
        format: uint8
        example: 16
        description: Minute:16/Hour:32/Day:48
      market:
        type: string
        example: abc/cet
...
  Redelegation:
    type: object
    properties:
      delegator:
        $ref: '#/definitions/Address'
      src:
        $ref: '#/definitions/Address'
      dst:
        $ref: '#/definitions/Address'
      amount:
        type: string
        example: "0"
      completion_time:
        type: string
        example: "0"
...